LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-empty-https-http-expected.txt

CONSOLE MESSAGE: The 'referrer' Content Security Policy directive has the invalid value "". Valid values are "no-referrer", "no-referrer-when-downgrade", "origin", "origin-when-cross-origin", and "unsafe-url".
Test that '' == never works correctly from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is ""
PASS successfullyParsed is true

TEST COMPLETE
